vmcleaner: move test helpers to main_test
