+++ go/src/koding/vmcleaner/warnings_test.go
@@ -188,22 +188,3 @@ func TestAct(t *testing.T) {
 		})
 	})
 }
-
-func createUser() (*models.User, error) {
-	username := "vminactive"
-	user := &models.User{
-		Name: username, ObjectId: bson.NewObjectId(),
-	}
-
-	return user, modelhelper.CreateUser(user)
-}
-
-func createInactiveUser(daysInactive int) (*models.User, error) {
-	user := &models.User{
-		Name: "inactiveuser", ObjectId: bson.NewObjectId(),
-		Email:         "inactiveuser@koding.com",
-		LastLoginDate: now().Add(-time.Hour * 24 * time.Duration(daysInactive)),
-	}
-
-	return user, modelhelper.CreateUser(user)
-}
